Move command logic into separate modules
